// src/api-client/models/register-recipient-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { RecipientType } from './recipient-type';

/**
 * 
 * @export
 * @interface RegisterRecipientDto
 */
export interface RegisterRecipientDto {
    /**
     * Sign-in address. Must not already be registered.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'email': string;
    /**
     * Plain-text password. Hashed before storage, never returned.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'password': string;
    /**
     * Name of the person registering. Becomes the primary contact for the organization and the signatory on its certificates.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'fullName': string;
    /**
     * Contact phone in E.164 format.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'phone'?: string;
    /**
     * The organization\'s public name.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'displayName': string;
    /**
     * What kind of receiver this is.
     * @type {RecipientType}
     * @memberof RegisterRecipientDto
     */
    'type': RecipientType;
    /**
     * Registered organization name, when it differs.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'legalName'?: string;
    /**
     * Government tax identifier. Must not already be registered when given.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'taxId'?: string;
    /**
     * Charity or non-profit registration number. Must not already be registered when given.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'registrationCode'?: string;
    /**
     * Where they operate, as they describe it.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'serviceArea'?: string;
    /**
     * IANA timezone. Pickup windows and the today/upcoming split are resolved in it.
     * @type {string}
     * @memberof RegisterRecipientDto
     */
    'timezone'?: string;
}