// src/api-client/models/register-retailer-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BusinessType } from './business-type';

/**
 * 
 * @export
 * @interface RegisterRetailerDto
 */
export interface RegisterRetailerDto {
    /**
     * Sign-in address. Must not already be registered.
     * @type {string}
     * @memberof RegisterRetailerDto
     */
    'email': string;
    /**
     * Plain-text password. Hashed before storage, never returned.
     * @type {string}
     * @memberof RegisterRetailerDto
     */
    'password': string;
    /**
     * Name of the person registering. Becomes the primary contact for the organization and the signatory on its certificates.
     * @type {string}
     * @memberof RegisterRetailerDto
     */
    'fullName': string;
    /**
     * Contact phone in E.164 format.
     * @type {string}
     * @memberof RegisterRetailerDto
     */
    'phone'?: string;
    /**
     * Registered company name.
     * @type {string}
     * @memberof RegisterRetailerDto
     */
    'legalName': string;
    /**
     * Public brand, when it differs from the legal name.
     * @type {string}
     * @memberof RegisterRetailerDto
     */
    'tradeName'?: string;
    /**
     * Government tax identifier (NIF/CIF/VAT). Must not already be registered.
     * @type {string}
     * @memberof RegisterRetailerDto
     */
    'taxId': string;
    /**
     * Kind of business, which shapes expected volume and logistics.
     * @type {BusinessType}
     * @memberof RegisterRetailerDto
     */
    'businessType': BusinessType;
}