// src/api-client/models/retailer-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BusinessType } from './business-type';
// May contain unused imports in some cases
// @ts-ignore
import type { ProfileStatus } from './profile-status';

/**
 * 
 * @export
 * @interface RetailerResponseDto
 */
export interface RetailerResponseDto {
    /**
     * Unique retailer ID.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'id': string;
    /**
     * Registered company name.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'legalName': string;
    /**
     * Public brand.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'tradeName'?: string | null;
    /**
     * Government tax identifier.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'taxId': string;
    /**
     * Kind of business.
     * @type {BusinessType}
     * @memberof RetailerResponseDto
     */
    'businessType': BusinessType;
    /**
     * Company description.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'description'?: string | null;
    /**
     * Public website.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'website'?: string | null;
    /**
     * Absolute URL of the logo.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'logoUrl'?: string | null;
    /**
     * Lifecycle status.
     * @type {ProfileStatus}
     * @memberof RetailerResponseDto
     */
    'status': ProfileStatus;
    /**
     * Whether an admin has verified this retailer.
     * @type {boolean}
     * @memberof RetailerResponseDto
     */
    'isVerified': boolean;
    /**
     * When the retailer was verified (ISO 8601).
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'verifiedAt'?: string | null;
    /**
     * ID of the admin who verified this retailer.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'verifiedById'?: string | null;
    /**
     * Food-safety licence number.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'foodSafetyLicenseNumber'?: string | null;
    /**
     * Licence expiry date.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'foodSafetyLicenseExpiresAt'?: string | null;
    /**
     * When the platform terms were accepted (ISO 8601).
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'termsAcceptedAt'?: string | null;
    /**
     * Version of the terms accepted.
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'termsVersion'?: string | null;
    /**
     * When the retailer was created (ISO 8601).
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'createdAt': string;
    /**
     * When the retailer was last modified (ISO 8601).
     * @type {string}
     * @memberof RetailerResponseDto
     */
    'updatedAt': string;
}