// src/api-client/models/user-role.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Authorization role.
 * @export
 * @enum {string}
 */

export const UserRole = {
    Admin: 'ADMIN',
    Retailer: 'RETAILER',
    Recipient: 'RECIPIENT'
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];